start changing preprocess to custom transformer.
