+++ automl/preprocessing/__init__.py
@@ -1,39 +1,47 @@
 # automl/preprocessing/__init__.py
 
-from .preprocessing import AutoML_Preprocess
+from .preprocessing import preprocess
 from .general import (
     drop_duplicate_rows,
     drop_duplicate_columns,
     drop_constant_columns,
     drop_strings,
+    skip_outliers,
+    detect_dataset_type,
 )
 from .outliers import (
-    skip_outliers,
     outlier_imputation_order,
     handle_outliers,
 )
-from .encoding import auto_encode_features, encode_target
-from .missing import handle_missing_values_num, handle_missing_values_cat
+from .encoding import auto_encode_features
+from .missing import handle_missing_values
 from .standardizing import (
     normalize_columns,
-    standardize_target,
+)
+from .target import (
+    is_target_categorical,
     TargetTransformer,
+    encode_target,
+    decode_target,
+    standardize_target,
 )
 
 __all__: list[str] = [
-    "AutoML_Preprocess",
+    "preprocess",
+    "is_target_categorical",
+    "decode_target",
     "drop_duplicate_rows",
     "drop_duplicate_columns",
     "drop_constant_columns",
     "skip_outliers",
     "outlier_imputation_order",
     "handle_outliers",
-    "handle_missing_values_cat",
-    "handle_missing_values_num",
+    "handle_missing_values",
     "auto_encode_features",
     "encode_target",
     "drop_strings",
     "normalize_columns",
     "standardize_target",
     "TargetTransformer",
+    "detect_dataset_type",
 ]
